Mapping Concepts to Terms Correctly Now
I am able to get this running now and it is mapping my concepts to terms correctly now. I have a couple of changes to may code that I want to try tomorrow. One is for embeddings of my words and the other one is to try and optimize memory.